fix: リネーム処理関連のリファクタリング.

+++ utils/rename_file_act_regular.py
@@ -3,6 +3,26 @@ import shutil
 import os
 
 
+# 同一拡張子のファイルが無いかチェックするプライベートメソッド
+def _check_duplicated_extends(filelist: list[str]) -> bool:
+    extends = []
+    duplicated_extends_count = 0
+
+    for file in filelist:
+        # os.path.splitext で確実に拡張子を取得する
+        extend = os.path.splitext(file)[1]
+        extends.append(extend)
+
+        for ext in extends:
+            if file.count(ext):
+                duplicated_extends_count += 1
+
+    if duplicated_extends_count > 0:
+        return True
+
+    return False
+
+
 # モード：all, part ルート時におけるリネーム処理
 def rename_file_act_regular(
     target_file_dir: list[str] | None = None,
@@ -23,8 +43,6 @@ def rename_file_act_regular(
         if os.path.isfile(f)
     ]
 
-    print(rename_files, target_file_dir, target_file_dir[0])
-
     if len(rename_files) == 0:
         print(f"{target_file_dir}内のファイルは現在「{len(rename_files)}」件です")
         return
@@ -37,24 +55,32 @@ def rename_file_act_regular(
         # part：ナンバリング有りver
         if is_add_numbering and target_str is not None:
             for i, file in enumerate(rename_files, 1):
+                # 置換対象文字を含んでいない場合はスキップ
+                if file.count(replace_str) == 0:
+                    print(f"{i} --- {file}は置換対象外です")
+                    continue
+
                 # バックアップを作成（.bak：バックアップファイルを意味する拡張子）
                 shutil.copy2(file, file + ".bak")
 
                 # 文字列置換
-                adjust_filename = file.replace(target_str, replace_str)
+                target_replace_str = replace_str
+                replace_result_str = target_str
+                adjust_filename = file.replace(target_replace_str, replace_result_str)
 
                 # ファイル名とディレクトリを分離
                 dir_name = os.path.dirname(adjust_filename)
                 base_name = os.path.basename(adjust_filename)
                 name_without_ext = os.path.splitext(base_name)[0]
-                extends = file.split(".")[-1]
+                # os.path.splitext で確実に拡張子を取得する
+                extend = os.path.splitext(base_name)[1]
 
                 # ファイル名の変更（`dir_name`と組み合わせてフルパス生成）
                 new_name = os.path.join(
                     dir_name,
-                    f"{i}-{name_without_ext}.{extends}"
+                    f"{i}-{name_without_ext}{extend}"
                     if numbering == "y"
-                    else f"{name_without_ext}-{i}.{extends}",
+                    else f"{name_without_ext}-{i}{extend}",
                 )
                 print(file, new_name)
                 os.rename(file, new_name)
@@ -63,8 +89,13 @@ def rename_file_act_regular(
         # part：ナンバリング無しver
         if target_str is not None:
             for file in rename_files:
+                if file.count(replace_str) == 0:
+                    continue
+
                 shutil.copy2(file, file + ".bak")
-                adjust_filename = file.replace(target_str, replace_str)
+                target_replace_str = replace_str
+                replace_result_str = target_str
+                adjust_filename = file.replace(target_replace_str, replace_result_str)
                 os.rename(file, adjust_filename)
             return
 
@@ -74,13 +105,13 @@ def rename_file_act_regular(
                 shutil.copy2(file, file + ".bak")
 
                 dir_name = os.path.dirname(file)
-                extends = file.split(".")[-1]
+                extend = os.path.splitext(file)[1]
 
                 new_name = os.path.join(
                     dir_name,
-                    f"{i}-{replace_str}.{extends}"
+                    f"{i}-{replace_str}{extend}"
                     if numbering == "y"
-                    else f"{replace_str}-{i}.{extends}",
+                    else f"{replace_str}-{i}{extend}",
                 )
                 print(file, new_name)
                 os.rename(file, new_name)
@@ -89,12 +120,18 @@ def rename_file_act_regular(
         # all：ナンバリング無しver
         else:
             for file in rename_files:
+                # 同一拡張子のファイルが無いかチェック
+                is_check_duplicated_extends = _check_duplicated_extends(rename_files)
+                if is_check_duplicated_extends:
+                    print("同じ拡張子のファイルがあります。重複ファイルは作成不可")
+                    return
+
                 shutil.copy2(file, file + ".bak")
 
                 dir_name = os.path.dirname(file)
-                extends = file.split(".")[-1]
+                extend = os.path.splitext(file)[1]
 
-                new_name = os.path.join(dir_name, f"{replace_str}.{extends}")
+                new_name = os.path.join(dir_name, f"{replace_str}{extend}")
                 os.rename(file, new_name)
             return
